dom/media/tests/mochitest/test_getUserMedia_basicWindowshare.html

<!DOCTYPE HTML>
<html>
<head>
  <script type="application/javascript" src="mediaStreamPlayback.js"></script>
</head>
<body>
<pre id="test">
<script type="application/javascript">
  createHTML({
    title: "getUserMedia Basic Windowshare Test",
    bug: "1038926"
  });
  /**
   * Run a test to verify that we can complete a start and stop media playback
   * cycle for an screenshare LocalMediaStream on a video HTMLMediaElement.
   */
  runTest(function () {
    const isWinXP = navigator.userAgent.indexOf("Windows NT 5.1") != -1;
    if (IsMacOSX10_6orOlder() || isWinXP) {
        ok(true, "Screensharing disabled for OSX10.6 and WinXP");
        SimpleTest.finish();
        return;
    }
    var testVideo = createMediaElement('video', 'testVideo');
    var constraints = {
      video: {
         mozMediaSource: "window",
         mediaSource: "window"
      },
      fake: false
    };

    getUserMedia(constraints).then(aStream => {
      checkMediaStreamTracks(constraints, aStream);

      var playback = new LocalMediaStreamPlayback(testVideo, aStream);
      return playback.playMediaWithStreamStop(false);
    }).then(() => SimpleTest.finish(), generateErrorCallback());

  });

</script>
</pre>
</body>
</html>
